chore: fix format warnings and update submodules
- Update Engine submodule to include dynamic_cast fixes
- Update UI submodule to include IUndoRedoCommand migration
- Fix UiConfigSelfCheck format specifier warning

+++ Main/Src/UI/ClientConfig/UiConfigSelfCheck.cpp
@@ -373,8 +373,8 @@ void UiConfigSelfCheck::logReport(const UiConfigSelfCheckReport& report, const Q
     }
     if (!report.gatedOutByBuild.isEmpty())
     {
-        SY_DEBUGF("[ConfigSelfCheck] %d command(s) absent because module not compiled (expected trim)",
-            report.gatedOutByBuild.size());
+        SY_DEBUGF("[ConfigSelfCheck] %lld command(s) absent because module not compiled (expected trim)",
+            static_cast<long long>(report.gatedOutByBuild.size()));
     }
 
     for (const QString& item : report.compiledButNotLicensed)
